+++ library/ic-ledger-types/CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.11.0] - 2024-05-17
+
+### Changed
+
+- Upgrade `ic-cdk` to v0.14.
+
 ## [0.10.0] - 2024-03-01
 
 ### Changed